requirements.txt
openai-whisper
deep-translator
streamlit
torch

